+++ CHANGES.md
@@ -6,7 +6,7 @@ Changes
 
 ### Unreleased
 
-* 2023-12-05 - Improvement: Option to suppress footer button, solves #444.
+* 2023-12-05 - Improvement: Option to suppress footer (circle containing the question mark) button, solves #444.
 * 2023-12-01 - Bugfix: Static pages unnecessarily cleaned configured content, solves #486.
 * 2023-11-23 - Bugfix: Add background color to OAuth2 login button, solves #473.